Fixes incorrect variable & property documentation; extract code to a function
The `_breakAfter` and `_failAfter` were both commented as being number of seconds. Actually, they are the number of write lock attempts to wait before trying to break the lock and failing the lock-breaking, respectively. So this PR updates the comments appropriately.

Also creates a function to return the number of write lock tries per second.

Renames _breakAfter and _failAfter properties to _breakAfterLockTries and _failAfterLockBreakTries to clarify what these values are.